+++ vulture/whitelists/logging_whitelist.py
@@ -1,4 +1,5 @@
 import logging
 
 logging.Filter.filter
+logging.getLogger().propagate
 logging.StreamHandler.emit